requirements.txt
# OCR Library Requirements
# Thư viện xử lý ảnh và OCR

# Core OCR engines
pytesseract>=0.3.10
easyocr>=1.7.0
python-doctr[torch]>=0.7.0
keras-ocr>=0.8.9

# Xử lý ảnh
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0

# Machine Learning (cho EasyOCR, DocTR và Keras OCR)
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0

# Tiện ích
matplotlib>=3.7.0
requests>=2.31.0
seaborn>=0.13.0
pandas>=1.5.0

# Optional: For GPU support (uncomment if you have CUDA)
# torch[cuda]
# torchvision[cuda]
# tensorflow[gpu]

# Development tools (optional)
jupyter>=1.0.0
pytest>=7.4.0